==> 04_accessibility.Rmd <==
---
title: "Logsums"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(cache = TRUE)
library(tidyverse)
library(sf)
library(spdep)
library(spatialreg)
library(leaflet)
library(lubridate)
library(VGAM)
```

In this document we calculate the accessiblity logsums based on the Euclidean
and network distances from the tract and block group centroids to the tract,
combined with park attributes.

```{r load}
tracts <- read_rds("data/tracts.rds")
open_spaces <- read_rds("data/open_spaces.rds")
distances <- read_rds("data/distances.rds")
```


## Distance Examination
The distances were computed according to the code presented in the [Distances](03_distances.html)
page. But we want to evaluate the difference between the various distance measurements
and visualize what they mean in terms of our analysis.

```{r read_distances}
blockgroup_files <- list.files("data/blockgroups", "*.csv", full.names = TRUE)
tract_files <- list.files("data/tracts", "*.csv", full.names = TRUE)

read_dist_csv <- function(x){
  # warnings are from `inf` strings, which become converted to NA, which is okay.
  suppressWarnings(read_csv(x, col_types = "ccnn") %>% as_tibble())
}

blockgroup_distances <- lapply(blockgroup_files, read_dist_csv) %>% 
  bind_rows() %>% 
  left_join(distances$blockgroups %>% rename(euc_old = distance), 
            by = c("geoid", "park_id"))
tract_distances <- lapply(tract_files, read_dist_csv) %>% 
  bind_rows() %>% 
  left_join(distances$tracts %>% rename(euc_old = distance), 
            by = c("geoid", "park_id"))
```

First, let's compare the difference between the Euclidean and network
distance at the block group level. Well, this is troubling; it appears as though
there is no relationship between the two methods of measurement. 

```{r bg_dist_comparison}
ggplot(blockgroup_distances %>% sample_n(1e4), aes(x = euc_dist, y = distance)) +
  geom_point(alpha = 0.3) + geom_smooth()
```

Let's look at the same plot for tracts. This is much better, though there
are still some curious issues. 

```{r tr_dist_comparison}
ggplot(tract_distances %>% sample_n(1e4), aes(x = euc_dist, y = distance)) +
  geom_point(alpha = 0.3) + geom_smooth()
```


In particular, how is it possible for the
Euclidean distance to be *longer* than the network distance? Let's look at this
by distance. It seems that everything is doing okay except for tracts and parks in Staten 
Island (prefix '085' and "R"). Obviously the shortest path isn't handling Staten
Island very well. The Verrazzano-Narrows Bridge is in the network, but I-78 and
the Bayonne Bridge route to Manhattan through New Jersey is not in the network.

```{r tr_dist_comparison_borough}
ggplot(tract_distances %>% sample_n(1e4) %>% 
         mutate(borough = substr(geoid, 3, 5),
                park_borough = substr(park_id, 0, 1),
                park_borough = ifelse(!park_borough %in% c("B", "M", "Q", "R", "X"), "C", park_borough)), 
       aes(x = euc_dist, y = distance)) +
  geom_abline(slope = 1, intercept = 0, lty = "dotted", color = "grey") + 
  geom_point(alpha = 0.3) + geom_smooth() +
  xlab("Euclidean Distance") + ylab("Network Distance") +
  facet_grid(borough ~ park_borough) +
  theme_bw()
```

All of this is to say we probably need to eliminate Staten Island for our 
analysis, or else adjust the network distances we calculate.

Is there a relation between the tract distance and the distances for block
groups within the tract?

```{r comparison}
comparison <- blockgroup_distances %>%
  sample_n(1e5) %>%
  rename(bg_net = distance, bg_euc = euc_dist) %>%
  mutate(tract = substr(geoid, 1, 11)) %>%
  left_join(tract_distances %>% rename(tr_net = distance, tr_euc = euc_dist), 
            by = c("tract" = "geoid", "park_id" = "park_id"))
```

Let's look first at the network distances. Yikes!
```{r comparison-net}
ggplot(comparison, aes(x = bg_net, y = tr_net)) +
  geom_point(alpha = 0.1) + geom_smooth() + 
  xlab("Blockgroup Network Distance") + ylab("Tract Network Distance")
```

The Euclidean distances make *loads* more sense, with some aggregation error
but it's obvious the numbers aren't stupid.

```{r comparison-net}
ggplot(comparison, aes(x = bg_euc, y = tr_euc)) +
  geom_point(alpha = 0.1) + geom_smooth() + 
  xlab("Blockgroup Euclidean Distance") + ylab("Tract Euclidean Distance")
```

Let's make a map of tracts colored by how far the two methods think they are
from Central Park. In general this seems fine, except again, Staten Island does
not look that great. Something is definitely happening there.

```{r distance-map}
pal <- colorQuantile(palette = "magma", log(tract_distances$distance), n = 7)
tracts %>%
  select(geoid) %>%
  left_join(tract_distances %>% filter(park_id == "M010"), by = "geoid") %>%
  st_transform(4326) %>%
  leaflet() %>%
  addProviderTiles(providers$Esri.WorldGrayCanvas) %>%
  addPolygons(color = ~pal(log(distance)), group = "Network") %>%
  addPolygons(color = ~pal(log(euc_dist)), group = "Euclidean") %>%
  addLayersControl(overlayGroups = c("Network", "Euclidean"))
```


## Logsums

Now that we have the distances between parks and block groups / tracts, we can 
calculate the logsums.

```{r logsums}
logsum <- function(x) { log(sum(exp(x))) }

compute_logsums <- function(df) {
  df %>%
    filter(!is.na(euc_old)) %>%
    left_join(open_spaces %>% st_set_geometry(NULL), by = c("park_id" = "id")) %>%
    # use euc_distance for impossible distances.
    mutate( distance = ifelse(distance < euc_dist, euc_dist, distance) ) %>%
    # convert distances to miles
    mutate_at(c("distance", "euc_dist"), ~ . * 0.000621371) %>%
    # must be at least 1/10 miles
    mutate_at(c("distance", "euc_dist"), ~ ifelse(. < 0.1, 0.1, .)) %>%
  
    # compute utility
    mutate(
      euc_access = -1.768 * log(euc_old) + 0.373 * log(as.numeric(acres)),
      euc_multi  = euc_access + trails * 0.99 +  playgrounds * 0.26 + courts * 0.43,
      net_access = -1.768 * log(distance) + 0.373 * log(as.numeric(acres)),
      net_multi  = net_access + trails * 0.99 +  playgrounds * 0.26 + courts * 0.43
    ) %>%
    
    # group by origin and compute access logsum
    group_by(geoid) %>%
    summarize_at(vars(euc_access:net_multi), logsum) %>%
    
    # standardize
    mutate_at(vars(euc_access:net_multi), ~(. - mean(.)) / sd(.))
    
}

tract_logsums <- compute_logsums(tract_distances) %>% write_rds("data/tract_logsums.rds")
bg_logsums    <- compute_logsums(blockgroup_distances) %>% write_rds("data/bg_logsums.rds")
```


Let's visualize the access logsums calculated for each graph in this method.

```{r tracts_access}
tracts_access <- tracts %>% left_join(tract_logsums, by = "geoid")
# this interactive map is much easier to create and explore
pal <- colorQuantile("GnBu", c(tracts_access$euc_access, tracts_access$net_multi), n = 7)
leaflet(tracts_access %>% st_transform(4326)) %>%
  addProviderTiles(providers$Esri.WorldGrayCanvas) %>%
  addPolygons(group = "Network", color = ~ pal(net_access),
              label = ~as.character(round(net_access, 2))) %>%
  addPolygons(group = "Network-Multi", color = ~ pal(net_multi),
              label = ~as.character(round(net_multi, 2))) %>%
  addPolygons(group = "Euclidean", color = ~ pal(euc_access),
              label = ~as.character(round(euc_access, 2))) %>%
  addPolygons(group = "Euclidean-Multi", color = ~ pal(euc_multi),
              label = ~as.character(round(euc_multi, 2))) %>%
  addLayersControl(baseGroups = c("Network", "Euclidean", "Network-Multi", "Euclidean-Multi"))
```


## 10-minute walk buffer

We are also interested in looking at what happens when the tracts are within a 
10-minute walk of a park, the threshold established by the Trust for Public Lands.

```{r nyc10mw}
access_10mw <- st_read("data/nyc_10mw.geojson") %>%
  st_cast("MULTIPOLYGON") %>%
  st_make_valid() %>%
  summarise(d = 1)

read_csv("data/tract_centroids.csv", col_types = "cnnnn") %>%
  st_as_sf(coords = c("LONGITUDE", "LATITUDE"), crs = 4326) %>%
  mutate(walk10 = st_intersects(., access_10mw, sparse = FALSE)) %>%
  st_set_geometry(NULL) %>%
  dplyr::select(geoid, walk10) %>%
  write_rds("data/walk10.rds")
```
